+++ CHANGELOG.md
@@ -2,13 +2,13 @@
 
 ## [next]
 
+- BREAKING refactor(Canvas): remove `button` property from mouse events
 - patch(Canvas): move event mouse:up:before earlier in the logic for more control [#9434](https://github.com/fabricjs/fabric.js/pull/9434)
 
 ## [6.0.0-beta15]
 
 - Fix(SVGParser) ignore missing xlink target issue on svg parsing (#9427) [#9109](https://github.com/fabricjs/fabric.js/issues/9109)
 - fix(#9172): dep export `Object`, `Text`, `Image` [#9433](https://github.com/fabricjs/fabric.js/pull/9433)
-- patch(Canvas): better respect of mouseup before [#9434](https://github.com/fabricjs/fabric.js/pull/9434)
 
 ## [6.0.0-beta14]